+++ src/main/java/org/michahirsch/hawkbit/ledstripe/ddi/client/script/RgbJson.java
@@ -5,31 +5,31 @@ package org.michahirsch.hawkbit.ledstripe.ddi.client.script;
 
 public class RgbJson {
 
-    private byte r;
-    private byte g;
-    private byte b;
+    private int r;
+    private int g;
+    private int b;
 
-    public byte getR() {
+    public int getR() {
         return r;
     }
 
-    public void setR(final byte r) {
+    public void setR(final int r) {
         this.r = r;
     }
 
-    public byte getG() {
+    public int getG() {
         return g;
     }
 
-    public void setG(final byte g) {
+    public void setG(final int g) {
         this.g = g;
     }
 
-    public byte getB() {
+    public int getB() {
         return b;
     }
 
-    public void setB(final byte b) {
+    public void setB(final int b) {
         this.b = b;
     }
 }
